Replaces GPU transform and data copy with GPU deformer node
Instead of hand-rolling a compute shader to transform vertices, then copy
data back to the CPU to set vertices on a mesh, this commit introduces
a GPU deformer node that handles the transformation directly on the GPU
(via native maya constructs).

The deformer takes time as an input to drive changes, which doesn't work
interactively while dragging during animation playback. To address this,
this commit also modifies the custom drag context to drive time manually
while dragging.

TODO:
- Cached playback support with the GPU deformer
- Get rid of the bind step altogether, bundle into GPU deformer.